@@ -26,39 +26,39 @@
     "dependencies": {
         "@discordjs/builders": "^0.8.2",
         "@discordjs/rest": "^0.1.0-canary.0",
+        "@prisma/client": "^3.5.0",
         "@sentry/integrations": "^6.15.0",
         "@sentry/node": "^6.15.0",
         "@sentry/tracing": "^6.15.0",
         "dayjs": "^1.10.7",
         "discord-api-types": "^0.24.0",
         "discord.js": "^13.3.1",
+        "dotenv": "^10.0.0",
         "fastify": "^3.24.0",
-        "i18next": "^21.5.2",
+        "i18next": "^21.5.3",
         "i18next-fs-backend": "^1.1.4",
         "is-image-url": "^1.1.8",
         "is-number": "^7.0.0",
-        "mongodb": "^4.2.0",
         "ms": "^2.1.3",
-        "node-cache": "^5.1.2",
         "node-schedule": "^2.0.0",
         "prom-client": "^14.0.1",
         "sha1": "^1.1.1",
         "tslib": "^2.3.1",
-        "undici": "^4.10.1",
-        "dotenv": "^10.0.0",
+        "undici": "^4.10.3",
         "winston": "^3.3.3"
     },
     "devDependencies": {
         "@types/i18next-fs-backend": "^1.1.2",
         "@types/is-number": "^7.0.1",
         "@types/ms": "^0.7.31",
-        "@types/node": "^16.11.9",
+        "@types/node": "^16.11.10",
         "@types/node-schedule": "^1.3.2",
         "@types/sha1": "^1.1.3",
         "@typescript-eslint/eslint-plugin": "^5.4.0",
         "@typescript-eslint/parser": "^5.4.0",
-        "eslint": "^8.2.0",
+        "eslint": "^8.3.0",
         "eslint-config-prettier": "^8.3.0",
+        "prisma": "^3.5.0",
         "standard-version": "^9.3.2",
         "typescript": "^4.5.2"
     },
